Update(mybulma): cantfix secvuln - issue #2
This vulnerability cannot presently and probably won't ever be fixed.

Node-sass is considered deprecated, which is the primary obstacle.

Even when using node-sass@6.0.0 on node LTS, the conflict remains.

Using `mybulma`, or any software with `node-sass` as a dependency is risky.

For more details, see the `npm-audit.txt` I added to this repo's root.